refactor: align tools presentation package with its directory
Keep Tools workspace and local tool dialogs in tools.presentation.
Shared scaffolds stay in ui.compose.
